Fabrice Kauffmann:
Here is the initial revision of the F-15c:

    http://fkauffmann.ifrance.com/archives/

I still need to find a solution for the strange surface behavior.
I noticed that other aircrafts have the same problem in FlightGear
(i.e. the F-16).  [ JSBSim flight control system ]

The reference points have to be corrected in the aero model too,
but I'm working on it.